portability: use uintptr_t for pointer arithmetic
Replaces (uint32_t) casts in pointer-difference and address-arithmetic
contexts with uintptr_t: string-offset math in the browser/load/save UIs, the
wavetable read in vector_rendering_function, and the element-key address math in
ordered_resizeable_array. These are the mechanical half of a 64-bit host build.

Firmware-neutral: on the 32-bit ARM target uintptr_t == uint32_t, so codegen is
byte-identical (firmware text 1694372 unchanged); the -m32 host-sim is likewise
unaffected (audio still bit-clean). What remains for a working 64-bit build is
not mechanical -- a sorted container that stores Drum* as 32-bit keys and casts
them back to pointers (kit.cpp), and the allocator's 32-bit-size assumptions --
so those are deliberately left for a later, scoped pass.

+++ src/deluge/processing/vector_rendering_function.h
@@ -33,7 +33,7 @@ waveRenderingFunctionGeneral(uint32_t phase, int32_t phaseIncrement, uint32_t _p
 		strength2[i] = rshifted;
 
 		uint32_t whichValue = phase >> (32 - tableSizeMagnitude);
-		uint32_t* readAddress = (uint32_t*)((uint32_t)table + (whichValue << 1));
+		uint32_t* readAddress = (uint32_t*)((uintptr_t)table + (whichValue << 1));
 		readValue[i].Load(readAddress);
 	});
